Move theme functions to themes.py
